Added tests, error handling for api, need last clean

+++ patient-api/src/main/java/com/mediscreen/patientapi/controller/PatientController.java
@@ -1,17 +1,19 @@
 package com.mediscreen.patientapi.controller;
 
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import javax.persistence.PersistenceException;
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.transaction.TransactionSystemException;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -27,10 +29,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mediscreen.patientapi.exception.UnknownDataException;
 import com.mediscreen.patientapi.model.Patient;
 import com.mediscreen.patientapi.service.PatientService;
-//TODO: renvoyer ResponnseEntity pour spécifier les codes HTTP
-//TODO: 
+//TODO: Revoir le nom racine 
 @RestController
 @RequestMapping("/patientapi")
+@CrossOrigin("${url.docnoteapi}")
 public class PatientController {
 
 	Logger logger = LoggerFactory.getLogger(PatientController.class);
@@ -44,55 +46,56 @@ public class PatientController {
 	@GetMapping("/get")
 	public JsonNode getPatient(@RequestParam Integer id, Model model) throws UnknownDataException {
 		
-		logger.info("GET /patient/get?id="+id);
+		logger.debug("GET /patient/get?id="+id);
 		Optional<Patient> opt = service.getPatient(id);
 		if (!opt.isPresent()) {
 			throw new UnknownDataException("This patient does not exist.");
 		}
 		Patient patient = opt.get();
 		model.addAttribute("patient",patient);
-		System.out.println("patient : "+patient);
+
 		return mapper.valueToTree(patient);
 	}
 	
 	@GetMapping("/getAll")
 	public JsonNode getAllPatients(Model model) {
-		logger.info("GET /patient/getAll");
+		logger.debug("GET /patient/getAll");
 		List<Patient> allPatients = service.getAllPatients();
 		model.addAttribute("patientlist", allPatients);
 		return mapper.valueToTree(allPatients);
 	}
-	
-	//TODO: vérifier que le patient n'existe pas déjà
-	//NOTE : OK!!
+	//TODO: pourquoi IllegalArgumentException
 	@PostMapping("/add")
 	public void addPatient(@RequestBody Patient patient) throws IllegalArgumentException{
-		logger.info("POST /patient/add?firstname="+patient.getFirstname()+"&lastname="+patient.getLastname()+"&birthdate="+patient.getBirthdate()+
+		logger.debug("POST /patient/add?firstname="+patient.getFirstname()+"&lastname="+patient.getLastname()+"&birthdate="+patient.getBirthdate()+
 				"&gender="+patient.getGender()+"&address="+patient.getAddress()+"&phone="+patient.getPhone());
 		
 		Patient saved = service.savePatient(patient);
-		logger.info("Success while adding patient(id="+saved.getId()+")");
+		logger.debug("Success while adding patient(id="+saved.getId()+")");
 	}
-
-	@PutMapping("/update/{id}")
-	public void updatePatient(@RequestBody Patient patient, @PathVariable Integer id) throws UnknownDataException,IllegalArgumentException {
+	
+	@PutMapping("/update")
+	public void updatePatient(@RequestBody Patient patient, @RequestParam Integer id) throws UnknownDataException,IllegalArgumentException {
 		
-		logger.info("PUT /patient/update/"+id+"?firstname="+patient.getFirstname()+"&lastname="+patient.getLastname()+"&birthdate="+patient.getBirthdate()+
+		logger.debug("PUT /patient/update/"+id+"?firstname="+patient.getFirstname()+"&lastname="+patient.getLastname()+"&birthdate="+patient.getBirthdate()+
 				"&gender="+patient.getGender()+"&address="+patient.getAddress()+"&phone="+patient.getPhone());
 		
 		Optional<Patient> opt = service.getPatient(id);
 		if (!opt.isPresent()) {
-			throw new UnknownDataException("This patient does not exist."); //TODO: Réecrire message partout !!
+			throw new UnknownDataException("This patient does not exist."); //FIXME: Réecrire message partout !!
 		}
 		patient.setId(id);
-		Patient updated = service.savePatient(patient);
-		logger.info("Success while updating patient(id="+updated.getId()+")");
+		
+		patient = service.savePatient(patient);
+		logger.debug("Success while updating patient : "+patient);
 	}
 	
-	@DeleteMapping("/delete/{id}")
-	public void deletePatient(@PathVariable("id") Integer id) {
-		
+	@DeleteMapping("/delete")
+	public void deletePatient(@RequestParam Integer id) throws UnknownDataException {
 		logger.info("DELETE /patient/delete/"+id);
-		service.deletePatient(id);
+		if(service.getPatient(id).isEmpty()) {
+			throw new UnknownDataException("This patient does not exist.");
+		}
+			service.deletePatient(id);
 	}
 }
